development-artifacts/research/requirements.txt
torch-directml>=0.2.0.dev230426
onnxruntime-directml>=1.16.0
celery>=5.3.4
redis>=5.0.1
transformers>=4.35.0
diffusers>=0.23.0
accelerate>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
pillow>=10.1.0
numpy>=1.24.0
safetensors>=0.4.0
huggingface-hub>=0.19.0
datasets>=2.14.0
tokenizers>=0.15.0
requests>=2.31.0
tqdm>=4.66.0
psutil>=5.9.0
GPUtil>=1.4.0
